refactor xai oauth extension into focused modules
